MAININTRO: split into parts.. done for 1 and 2
